removed remaning pulse code. added support for doubles instead of bytes. minus unary operator now works

+++ Evoke/Evoke/environment.cpp
@@ -40,7 +40,7 @@ Value Environment::getVariable(Token name) const
 		return values.at(name.lexeme);
 	}
 	throw RuntimeError(name, "Undefined variable '" + name.lexeme + "'.");
-	return byte();
+	return double();
 }
 
 void Environment::defineArray(const std::string& name)
